renaming to image-to-text
