+++ widgets/src/lib/InferenceWidget/widgets/ImageToTextWidget/ImageToTextWidget.svelte
